# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Najjača aplikacija u galaksiji API!",
    "description": "Ovo je najjača aplikacija u galaksiji i šire. Onaj koji ju koristi je najjači u galaksiji! (p.s. Sandi je kriv za ovu tvorevinu:))",
    "version": "v1"
  },
  "paths": {
    "/application/paged": {
      "post": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/application": {
      "get": {
        "tags": [
          "Application"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/application/{id}": {
      "get": {
        "tags": [
          "Application"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Application"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/nationalidtype/paged": {
      "post": {
        "tags": [
          "NationalIdType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/nationalidtype": {
      "get": {
        "tags": [
          "NationalIdType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "NationalIdType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "NationalIdType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NationalIdTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/nationalidtype/{id}": {
      "get": {
        "tags": [
          "NationalIdType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "NationalIdType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/place/paged": {
      "post": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TableMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/place": {
      "get": {
        "tags": [
          "Place"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/place/{id}": {
      "get": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationDto": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "applicationName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FilterMetadata": {
        "type": "object",
        "properties": {
          "filterColumn": {
            "type": "string",
            "nullable": true
          },
          "filterValue": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NationalIdTypeDto": {
        "type": "object",
        "properties": {
          "nationalIdTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "nationalIdTypeName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PagingMetadata": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PlaceDto": {
        "type": "object",
        "properties": {
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "placeName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SortMetadata": {
        "type": "object",
        "properties": {
          "orderBy": {
            "type": "string",
            "nullable": true
          },
          "orderDirection": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TableMetadata": {
        "type": "object",
        "properties": {
          "filterMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterMetadata"
            },
            "nullable": true
          },
          "sortMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortMetadata"
            },
            "nullable": true
          },
          "pagingMetadata": {
            "$ref": "#/components/schemas/PagingMetadata"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Specify API Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}